sql/ast: Export ast.go into individual files
