+++ internal/sql/ast/alter_table_set_schema_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type AlterTableSetSchemaStmt struct {
+	Table     *TableName
+	NewSchema *string
+}
+
+func (n *AlterTableSetSchemaStmt) Pos() int {
+	return 0
+}
+
+
