+++ internal/sql/ast/alter_table_stmt.go
@@ -1,0 +1,11 @@
+package ast
+
+type AlterTableStmt struct {
+	Table *TableName
+	Cmds  *List
+	// MissingOk bool
+}
+
+func (n *AlterTableStmt) Pos() int {
+	return 0
+}
