+++ internal/sql/ast/alter_type_add_value_stmt.go
@@ -1,0 +1,13 @@
+package sqlc
+
+type AlterTypeAddValueStmt struct {
+	Type               *TypeName
+	NewValue           *string
+	SkipIfNewValExists bool
+}
+
+func (n *AlterTypeAddValueStmt) Pos() int {
+	return 0
+}
+
+
