+++ internal/sql/ast/alter_type_rename_value_stmt.go
@@ -1,0 +1,11 @@
+package sqlc
+
+type AlterTypeRenameValueStmt struct {
+	Type     *TypeName
+	OldValue *string
+	NewValue *string
+}
+
+func (n *AlterTypeRenameValueStmt) Pos() int {
+	return 0
+}
