+++ internal/sql/ast/column_def.go
@@ -1,0 +1,14 @@
+package sqlc
+
+// TODO: Support array types
+type ColumnDef struct {
+	Colname   string
+	TypeName  *TypeName
+	IsNotNull bool
+}
+
+func (n *ColumnDef) Pos() int {
+	return 0
+}
+
+
