+++ internal/sql/ast/column_ref.go
@@ -1,0 +1,9 @@
+package sqlc
+
+type ColumnRef struct {
+	Name string
+}
+
+func (n *ColumnRef) Pos() int {
+	return 0
+}
