+++ internal/sql/ast/comment_on_column_stmt.go
@@ -1,0 +1,13 @@
+package sqlc
+
+type CommentOnColumnStmt struct {
+	Table   *TableName
+	Col     *ColumnRef
+	Comment *string
+}
+
+func (n *CommentOnColumnStmt) Pos() int {
+	return 0
+}
+
+
