+++ internal/sql/ast/comment_on_schema_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type CommentOnSchemaStmt struct {
+	Schema  *String
+	Comment *string
+}
+
+func (n *CommentOnSchemaStmt) Pos() int {
+	return 0
+}
+
+
