+++ internal/sql/ast/comment_on_table_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type CommentOnTableStmt struct {
+	Table   *TableName
+	Comment *string
+}
+
+func (n *CommentOnTableStmt) Pos() int {
+	return 0
+}
+
+
