+++ internal/sql/ast/comment_on_type_stmt.go
@@ -1,0 +1,10 @@
+package sqlc
+
+type CommentOnTypeStmt struct {
+	Type    *TypeName
+	Comment *string
+}
+
+func (n *CommentOnTypeStmt) Pos() int {
+	return 0
+}
