+++ internal/sql/ast/create_enum_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type CreateEnumStmt struct {
+	TypeName *TypeName
+	Vals     *List
+}
+
+func (n *CreateEnumStmt) Pos() int {
+	return 0
+}
+
+
