+++ internal/sql/ast/create_function_stmt.go
@@ -1,0 +1,14 @@
+package sqlc
+
+type CreateFunctionStmt struct {
+	Replace    bool
+	Params     []*FuncParam
+	ReturnType *TypeName
+	Func       *FuncName
+}
+
+func (n *CreateFunctionStmt) Pos() int {
+	return 0
+}
+
+
