+++ internal/sql/ast/create_schema_stmt.go
@@ -1,0 +1,10 @@
+package sqlc
+
+type CreateSchemaStmt struct {
+	Name        *string
+	IfNotExists bool
+}
+
+func (n *CreateSchemaStmt) Pos() int {
+	return 0
+}
