+++ internal/sql/ast/create_table_stmt.go
@@ -1,0 +1,13 @@
+package sqlc
+
+type CreateTableStmt struct {
+	IfNotExists bool
+	Name        *TableName
+	Cols        []*ColumnDef
+}
+
+func (n *CreateTableStmt) Pos() int {
+	return 0
+}
+
+
