+++ internal/sql/ast/drop_function_stmt.go
@@ -1,0 +1,10 @@
+package sqlc
+
+type DropFunctionStmt struct {
+	Funcs     []*FuncSpec
+	MissingOk bool
+}
+
+func (n *DropFunctionStmt) Pos() int {
+	return 0
+}
