+++ internal/sql/ast/drop_schema_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type DropSchemaStmt struct {
+	Schemas   []*String
+	MissingOk bool
+}
+
+func (n *DropSchemaStmt) Pos() int {
+	return 0
+}
+
+
