+++ internal/sql/ast/drop_table_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type DropTableStmt struct {
+	IfExists bool
+	Tables   []*TableName
+}
+
+func (n *DropTableStmt) Pos() int {
+	return 0
+}
+
+
