+++ internal/sql/ast/drop_type_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type DropTypeStmt struct {
+	IfExists bool
+	Types    []*TypeName
+}
+
+func (n *DropTypeStmt) Pos() int {
+	return 0
+}
+
+
