+++ internal/sql/ast/func_name.go
@@ -1,0 +1,13 @@
+package sqlc
+
+type FuncName struct {
+	Catalog string
+	Schema  string
+	Name    string
+}
+
+func (n *FuncName) Pos() int {
+	return 0
+}
+
+
