+++ internal/sql/ast/func_param.go
@@ -1,0 +1,11 @@
+package sqlc
+
+type FuncParam struct {
+	Name    *string
+	Type    *TypeName
+	DefExpr Node // Will always be &ast.TODO
+}
+
+func (n *FuncParam) Pos() int {
+	return 0
+}
