+++ internal/sql/ast/func_spec.go
@@ -1,0 +1,11 @@
+package sqlc
+
+type FuncSpec struct {
+	Name    *FuncName
+	Args    []*TypeName
+	HasArgs bool
+}
+
+func (n *FuncSpec) Pos() int {
+	return 0
+}
