+++ internal/sql/ast/list.go
@@ -1,0 +1,9 @@
+package sqlc
+
+type List struct {
+	Items []Node
+}
+
+func (n *List) Pos() int {
+	return 0
+}
