+++ internal/sql/ast/node.go
@@ -1,0 +1,5 @@
+package ast
+
+type Node interface {
+	Pos() int
+}
