+++ internal/sql/ast/raw_stmt.go
@@ -1,0 +1,11 @@
+package ast
+
+type RawStmt struct {
+	Stmt         Node
+	StmtLocation int
+	StmtLen      int
+}
+
+func (n *RawStmt) Pos() int {
+	return n.StmtLocation
+}
