+++ internal/sql/ast/rename_column_stmt.go
@@ -1,0 +1,13 @@
+package sqlc
+
+type RenameColumnStmt struct {
+	Table   *TableName
+	Col     *ColumnRef
+	NewName *string
+}
+
+func (n *RenameColumnStmt) Pos() int {
+	return 0
+}
+
+
