+++ internal/sql/ast/rename_table_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type RenameTableStmt struct {
+	Table   *TableName
+	NewName *string
+}
+
+func (n *RenameTableStmt) Pos() int {
+	return 0
+}
+
+
