+++ internal/sql/ast/res_target.go
@@ -1,0 +1,11 @@
+package sqlc
+
+type ResTarget struct {
+	Val Node
+}
+
+func (n *ResTarget) Pos() int {
+	return 0
+}
+
+
