+++ internal/sql/ast/select_stmt.go
@@ -1,0 +1,12 @@
+package sqlc
+
+type SelectStmt struct {
+	Fields *List
+	From   *List
+}
+
+func (n *SelectStmt) Pos() int {
+	return 0
+}
+
+
