+++ internal/sql/ast/statement.go
@@ -1,0 +1,9 @@
+package ast
+
+type Statement struct {
+	Raw *RawStmt
+}
+
+func (n *Statement) Pos() int {
+	return 0
+}
