+++ internal/sql/ast/string.go
@@ -1,0 +1,9 @@
+package sqlc
+
+type String struct {
+	Str string
+}
+
+func (n *String) Pos() int {
+	return 0
+}
