+++ internal/sql/ast/table_name.go
@@ -1,0 +1,11 @@
+package ast
+
+type TableName struct {
+	Catalog string
+	Schema  string
+	Name    string
+}
+
+func (n *TableName) Pos() int {
+	return 0
+}
