+++ internal/sql/ast/todo.go
@@ -1,0 +1,8 @@
+package sqlc
+
+type TODO struct {
+}
+
+func (n *TODO) Pos() int {
+	return 0
+}
