+++ internal/sql/ast/type_name.go
@@ -1,0 +1,13 @@
+package sqlc
+
+type TypeName struct {
+	Catalog string
+	Schema  string
+	Name    string
+}
+
+func (n *TypeName) Pos() int {
+	return 0
+}
+
+
